file added

+++ variable_datatype/p50.java
@@ -1,6 +1,6 @@
 class demo{
 	public static void main(String [] args){
-		float a=10.5;
+		int a=10.5f;
 		System.out.println(a);	
 	}
 }
