+++ variable_datatype/p51.java
@@ -1,6 +1,6 @@
 class demo{
 	public static void main(String [] args){
-		float a=10.5;
+		double a=10.5f;
 		System.out.println(a);	
 	}
 }
